feat: validate user to change hunts data

+++ src/modules/hunt/controllers/hunt-collection.controller.ts
@@ -11,6 +11,7 @@ import {
   Post,
   Put,
   Query,
+  UnauthorizedException,
   UseGuards,
   UseInterceptors
 } from '@nestjs/common'
@@ -113,7 +114,10 @@ export class HuntController {
   @ApiBearerAuth()
   @UseGuards(AuthGuard)
   @Get(':id')
-  async getOneHuntById(@Param('id') id: string) {
+  async getOneHuntById(
+    @Param('id') id: string,
+    @CurrentUser() user: AuthenticatedUser
+  ) {
     if (!id) {
       throw new BadRequestException('É necessário informar a id da hunt')
     }
@@ -122,6 +126,12 @@ export class HuntController {
 
     if (!found) {
       throw new NotFoundException('Hunt não encontrada')
+    }
+
+    const validated = await this.huntService.validateUserAccess(user.id, id)
+
+    if (!validated) {
+      throw new UnauthorizedException('Usuário sem permissão nesta hunt')
     }
 
     return found
@@ -143,7 +153,8 @@ export class HuntController {
   async updateHunt(
     @Param('id') id: string,
     @Body(new ZodValidationPipe(updateHuntSchema))
-    updateData: UpdateHunt
+    updateData: UpdateHunt,
+    @CurrentUser() user: AuthenticatedUser
   ) {
     if (!id) {
       throw new BadRequestException('É necessário informar a id da hunt')
@@ -153,6 +164,12 @@ export class HuntController {
 
     if (!found) {
       throw new NotFoundException('A hunt não existe')
+    }
+
+    const validated = await this.huntService.validateUserAccess(user.id, id)
+
+    if (!validated) {
+      throw new UnauthorizedException('Usuário sem permissão nesta hunt')
     }
 
     return await this.huntService.updateHunt(id, {
@@ -187,7 +204,10 @@ export class HuntController {
   @ApiBearerAuth()
   @UseGuards(AuthGuard)
   @Delete(':id')
-  async deleteHunt(@Param('id') id: string) {
+  async deleteHunt(
+    @Param('id') id: string,
+    @CurrentUser() user: AuthenticatedUser
+  ) {
     if (!id) {
       throw new BadRequestException('É necessário informar a id da hunt')
     }
@@ -196,6 +216,12 @@ export class HuntController {
 
     if (!toDelete) {
       throw new NotFoundException('Hunt não encontrada')
+    }
+
+    const validated = await this.huntService.validateUserAccess(user.id, id)
+
+    if (!validated) {
+      throw new UnauthorizedException('Usuário sem permissão nesta hunt')
     }
 
     const deleted = await this.huntService.deleteHunt(id)
